IMPALA-9180 (part 2): Refactor executor_list_ map of ExecuterBlacklist
In current ExecuterBlacklist class, we've keyed maps on a
TNetworkAddress of a backend. To simply the logic for the class,
changes it to key off of the UniqueIdPB backend-id, eg. refactor
'executor_list_' to no longer be a map<address -> list<backend>>
and instead makes it a map<backend_id, backend>.
Also fixes a minor bug with the calculation of elapsed time when
a backend that was on probation is re-blacklisted.

Testing:
 - Passed test_blacklist.py and test_query_retries.py.
 - Passed exhaustive tests.

Change-Id: Ib1ae082d0e080088756af91b5b770752ca8b3aa1
Reviewed-on: http://gerrit.cloudera.org:8080/16506
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/105657
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Sahil Takiar <stakiar@cloudera.com>