IMPALA-10143: TestAcid.test_full_acid_original_files is flaky
One of the test queries depended on the values of the row__id.rowid
field. In rare cases the files are written differently by Hive which
leads to different 'rowid' values for the rows.

I took out the test for the 'rowid' values in that particular query.

We have other tests for the 'rowid' field on static data files (from
'testdata/data'), therefore we still have coverage for that and they
shouldn't be flaky.

Change-Id: I3d36bd23b8d3cc257bad9a83a4462f20e073d437
Reviewed-on: http://gerrit.cloudera.org:8080/16523
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/105644
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Sahil Takiar <stakiar@cloudera.com>