IMPALA-9664: Fix typo in test_event_processing.py
The test tries to add src_db within a database
object which would fail when the database
managed location is present. This test doesn't
fail currently in ASF master since the database
doesn't have managed location yet. It will start
failing once the managed location for databases is
available in the toolchain build of Hive.

Testing:
1. The test was working before the patch since the
managed db location was probably not set and modified
line was not getting executed. I made sure the test works
with the patch as well.
2. I applied the patch in an environment where
managed db location is available and the error disappears.
(Although the test fails for another unrelated reason
HIVE-23995) so we should be aware of this when the toolchain
hive build is bumped up.

Change-Id: I34e16f52722ada2334aeb3bbb187c6c6358d20a3
Reviewed-on: http://gerrit.cloudera.org:8080/16547
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/105656
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Sahil Takiar <stakiar@cloudera.com>