IMPALA-10383: Fix data race on AdmissionController::queue_nodes_
IMPALA-9930 introduces a new data structure to AdmissionController:
'queue_nodes_'. The intention was that access to 'queue_nodes_' should
always be protected by 'queue_nodes_lock_', but this was not always
the case.

This patch fixes the issue by ensuring that 'queue_nodes_lock_' is
held everywhere that 'queue_nodes_' is used.

Testing:
- Ran TSAN and confirmed there is no longer a data race being
  reported.

Change-Id: I05b70112300d3a886bdd5f02fda0bd7ca98df6cf